requirements.txt
flask
flask-cors
paho-mqtt
torch
torchvision
scikit-learn
pillow
joblib
gunicorn
numpy
